+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/jdkversiontests/java22/Jep456_UnnamedPatternsAndVariables.java
@@ -9,9 +9,10 @@ import java.util.Queue;
 import java.util.stream.Collectors;
 
 /**
- * @see <a href="https://openjdk.org/jeps/443">JEP 443: Unnamed Patterns and Variables (Preview)</a>
+ * @see <a href="https://openjdk.org/jeps/443">JEP 443: Unnamed Patterns and Variables (Preview)</a> (Java 21)
+ * @see <a href="https://openjdk.org/jeps/456">JEP 456: Unnamed Variables & Patterns</a> (Java 22)
  */
-class Jep443_UnamedPatternsAndVariables {
+class Jep456_UnamedPatternsAndVariables {
     record Point(int x, int y) { }
     enum Color { RED, GREEN, BLUE }
     record ColoredPoint(Point p, Color c) { }